Exercise 6.2

+++ framework/shape.hpp
@@ -7,8 +7,8 @@ class Shape
 			Shape();
 			virtual ~Shape();
 
-			virtual float volume() const = 0;
-			virtual float area() const = 0;
+			virtual double volume() const = 0;
+			virtual double area() const = 0;
 
 };
